Edited lendor.dashboard.lender_portfolio_first_page.RowTemplate32 and lendor.dashboard.lender_portfolio_first_page; Deleted borrower.dashboard.borrower_portfolio_first_page.RowTemplate31
