+++ client_code/lendor/dashboard/lender_portfolio_first_page/RowTemplate32/form_template.yaml
@@ -1,3 +1,32 @@
 is_package: true
 container: {type: DataRowPanel}
-components: []
+components:
+- name: label_1
+  properties: {align: center}
+  type: Label
+  layout_properties: {column: JKSDUP}
+  data_bindings:
+  - {property: text, code: 'self.item[''name'']', writeback: false}
+- name: label_2
+  properties: {align: center}
+  type: Label
+  layout_properties: {column: VJQQOR}
+  data_bindings:
+  - {property: text, code: 'self.item[''email'']', writeback: false}
+- name: label_3
+  properties: {align: center}
+  type: Label
+  layout_properties: {column: OJLQEN}
+  data_bindings:
+  - {property: text, code: 'self.item[''lending_type'']', writeback: false}
+- name: label_4
+  properties: {align: center}
+  type: Label
+  layout_properties: {column: RBCPXE}
+  data_bindings:
+  - {property: text, code: 'self.item[''membership'']', writeback: false}
+- name: link_1
+  properties: {text: View Portfolio, visible: false, align: center}
+  type: Link
+  layout_properties: {column: GONDTU}
+  event_bindings: {click: link_1_click}
